+++ test/ReadonlyArray.ts
@@ -1398,4 +1398,9 @@ describe('ReadonlyArray', () => {
     U.deepStrictEqual(_.fromOption(O.some('hello')), ['hello'])
     U.deepStrictEqual(_.fromOption(O.none), [])
   })
+
+  it('fromEither', () => {
+    U.deepStrictEqual(_.fromEither(E.right(1)), [1])
+    U.strictEqual(_.fromEither(E.left('a')), _.empty)
+  })
 })