+++ CHANGELOG.md
@@ -60,6 +60,9 @@ high state of flux, you're at risk of it changing without notice.
     - add `getUnionMonoid`
     - add `getIntersectionSemigroup`
     - add `getDifferenceMagma`
+    - add `fromEither`
+    - add `FromEither` instance
+    - add `fromEitherK`
   - `Either`
     - add `chainOptionK`
   - `function`
@@ -87,6 +90,9 @@ high state of flux, you're at risk of it changing without notice.
     - add `getDifferenceMagma`
     - add `fromOption`
     - add `fromPredicate`
+    - add `fromEither`
+    - add `FromEither` instance
+    - add `fromEitherK`
   - `ReadonlyNonEmptyArray`
     - add `matchLeft`, `matchRight`, `modifyHead`, `modifyLast` (@cdimitroulas)
     - add `union`